// models/AddFieldsData.ts
/**
 * DocSpring API
 * DocSpring provides an API that helps you fill out and sign PDF templates.
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class AddFieldsData {
    'fields': Array<any>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "fields",
            "baseName": "fields",
            "type": "Array<any>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AddFieldsData.attributeTypeMap;
    }

    public constructor() {
    }
}